examples/simple-fastapi-demo/requirements.txt
gino-starlette==0.1.0a0
fastapi
uvicorn==0.11.2
